stylenashky/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-import-export==2.7.1
et-xmlfile==1.1.0
idna==3.3
requests==2.27.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
xlrd==2.0.1
urllib3==1.26.8
xlwt==1.3.0
